ReactJS: Move and rename the proxy adapter class.
