+++ views-react/src/test/groovy/io/micronaut/views/react/IntrospectableBeansAreProxiedSpec.groovy
@@ -1,6 +1,7 @@
 package io.micronaut.views.react
 
 import io.micronaut.test.extensions.spock.annotation.MicronautTest
+import io.micronaut.views.react.truffle.IntrospectableToTruffleAdapter
 import io.micronaut.views.react.util.BeanPool
 import jakarta.inject.Inject
 import org.graalvm.polyglot.Value
@@ -19,7 +20,7 @@ class IntrospectableBeansAreProxiedSpec extends Specification {
         def bean = new SomeBean("foo value", "bar value", new SomeBean.InnerBean(10, Map.of("key", 123), List.of("one", "two", "three")))
 
         when:
-        ProxyObject proxy = ProxyObjectWithIntrospectableSupport.wrap(context, bean).asProxyObject()
+        ProxyObject proxy = IntrospectableToTruffleAdapter.wrap(context, bean).asProxyObject()
         context.getBindings("js").putMember("bean", proxy)
 
         then:
@@ -30,7 +31,7 @@ class IntrospectableBeansAreProxiedSpec extends Specification {
         ProxyObject innerBean = ((Value) proxy.getMember("innerBean")).asProxyObject()
 
         then:
-        innerBean instanceof ProxyObjectWithIntrospectableSupport
+        innerBean instanceof IntrospectableToTruffleAdapter
 
         when:
         ProxyObject innerBeanMap = ((Value) innerBean.getMember("map")).asProxyObject()
